Add new metaball edge advanced screen
